+++ tests/testthat/test-is.R
@@ -29,6 +29,17 @@ with_dir_tree(list("foo/bar"  = "test"), {
       expect_equal(is_link("baz"), c(baz = FALSE))
     })
   })
+
+  describe("is_link", {
+    it("returns true for empty files and false for others", {
+      file_create("blah")
+      expect_true(is_empty("blah"))
+      expect_false(is_empty("foo"))
+      expect_false(is_empty("foo/bar"))
+      expect_false(is_empty("baz"))
+    })
+  })
+
   describe("is_absolute_path", {
     it("detects windows absolute paths", {
       expect_true(is_absolute_path("c:\\"))